Do not enable GeoIP serverwide by default.
This is not recommended by MaxMind and also not logical: a server may have
many vhosts and only a small subset of hosts or paths may actually need
the GeoIP information in the request. Therefore, it's much better to
enable it selectively for those vhosts/locations/directories where it
will actually be used.